@@ -41,18 +41,19 @@
     "@ephox/bedrock-server": "^11.3.1",
     "@ephox/katamari": "^7.2.1",
     "@ephox/sugar": "^7.1.3",
-    "@storybook/addon-actions": "^6.4.0-alpha.9",
-    "@storybook/addon-essentials": "^6.4.0-alpha.9",
-    "@storybook/addon-links": "^6.4.0-alpha.9",
+    "@storybook/addon-actions": "^6.4.18",
+    "@storybook/addon-essentials": "^6.4.18",
+    "@storybook/addon-links": "^6.4.18",
     "@storybook/addon-notes": "^5.3.21",
     "@storybook/storybook-deployer": "^2.8.10",
-    "@storybook/vue3": "^6.4.0-alpha.9",
+    "@storybook/vue3": "^6.4.18",
     "@tinymce/beehive-flow": "^0.17.0",
     "@tinymce/eslint-plugin": "^1.9.1",
     "@tinymce/miniature": "^3.2.0",
     "@types/node": "^16.7.1",
-    "@vue/compiler-sfc": "^3.0.0",
-    "babel-loader": "^8.0.5",
+    "@vitejs/plugin-vue": "^2.1.0",
+    "@vue/compiler-sfc": "^3.2.29",
+    "babel-loader": "^8.2.3",
     "babel-preset-vue": "^2.0.2",
     "babel-register": "^6.26.0",
     "css-loader": "^6.2.0",
@@ -67,12 +68,13 @@
     "rollup-plugin-uglify": "^6.0.0",
     "tinymce-4": "npm:tinymce@^4",
     "tinymce-5": "npm:tinymce@^5",
-    "ts-loader": "^9.2.5",
+    "ts-loader": "^9.2.6",
     "ts-node": "^10.2.1",
-    "tsconfig-paths-webpack-plugin": "^3.5.1",
+    "tsconfig-paths-webpack-plugin": "^3.5.2",
     "typescript": "^4.3.5",
-    "vite": "^2.5.1",
-    "vue": "^3.0.0",
+    "vite": "^2.7.13",
+    "vue": "^3.2.5",
+    "vue-cli-plugin-storybook": "^2.1.0",
     "vue-loader": "^15.2.1",
     "vue-router": "^4.0.0-beta.13",
     "vue-template-compiler": "^2.6.14",
