@@ -6,6 +6,8 @@
       <router-link :class="{ active: $route.path==='/inline'}" to="/inline">Inline Editor</router-link>
       <router-link :class="{ active: $route.path==='/controlled'}" to="/controlled">Controlled component</router-link>
       <router-link :class="{ active: $route.path==='/keepalive'}" to="/keepalive">Keep-alive</router-link>
+      <router-link :class="{ active: $route.path==='/refreshable'}" to="/refreshable">Rerender</router-link>
+      <router-link :class="{ active: $route.path==='/tagged'}" to="/tagged">Tag Change</router-link>
     </div>
     <router-view></router-view>
   </div>
@@ -15,5 +17,5 @@
 import { defineComponent } from "vue";
 export default defineComponent({
   name: 'Demo',
-})
+});
 </script>
